Python 3 support, fixed two typos
So, as above; Added Python 3 compatibility (Python 2.7 still works, of course), fixed two typos. The funny fact is that I bricked my phone during this PR. I wanted to check if this works, but it seem like my wifi won't turn out. Then I find out that yesterday, I have compiled kernel with Wi-Fi bug, so I went straight to xda for some patch, but I confused versions and this resulted in soft brick. Fortunately, I just flashed newest firmware and it works now. This story probably should not be in this PR, but why not. Nobody reads PR descriptions.